Add vhd writing support
This adds a VHDUtilImageManager class that supports writing out tar
files that contain vhds. This relies on a 3rd party tool, 'vhd-util',
and patches to that utility.

A script to build vhd-util exists here:

https://github.com/comstud/libvhd-builder

That will clone the appropriate repo for vhd-util and build it, as long
as you have the correct dependencies (documented in the build script).

Change-Id: I24d7940040e5edc413469b434778409bbac0c8f3